+++ x/vesting/keeper/utils_test.go
@@ -1,0 +1,90 @@
+package keeper_test
+
+import (
+	"strings"
+
+	. "github.com/onsi/gomega"
+
+	"cosmossdk.io/math"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	"github.com/evmos/evmos/v11/testutil"
+
+	cosmosante "github.com/evmos/evmos/v11/app/ante/cosmos"
+	evmante "github.com/evmos/evmos/v11/app/ante/evm"
+	"github.com/evmos/evmos/v11/utils"
+	"github.com/evmos/evmos/v11/x/vesting/types"
+)
+
+// assertEthFails is a helper function that takes in 1 or more messages and checks
+// that they can neither be validated nor delivered using the EthVesting.
+func assertEthFails(msgs ...sdk.Msg) {
+	insufficientUnlocked := "insufficient unlocked"
+
+	err := validateEthVestingTransactionDecorator(msgs...)
+	Expect(err).ToNot(BeNil())
+	Expect(strings.Contains(err.Error(), insufficientUnlocked))
+
+	// Sanity check that delivery fails as well
+	_, err = testutil.DeliverEthTx(s.ctx, s.app, nil, msgs...)
+	Expect(err).ToNot(BeNil())
+	Expect(strings.Contains(err.Error(), insufficientUnlocked))
+}
+
+// assertEthSucceeds is a helper function, that checks if 1 or more messages
+// can be validated and delivered.
+func assertEthSucceeds(testAccounts []TestClawbackAccount, funder sdk.AccAddress, dest sdk.AccAddress, amount math.Int, denom string, msgs ...sdk.Msg) {
+	numTestAccounts := len(testAccounts)
+
+	// Track starting balances for all accounts
+	granteeBalances := make(sdk.Coins, numTestAccounts)
+	funderBalance := s.app.BankKeeper.GetBalance(s.ctx, funder, denom)
+	destBalance := s.app.BankKeeper.GetBalance(s.ctx, dest, denom)
+
+	for i, grantee := range testAccounts {
+		granteeBalances[i] = s.app.BankKeeper.GetBalance(s.ctx, grantee.address, denom)
+	}
+
+	// Validate the AnteHandler passes without issue
+	err := validateEthVestingTransactionDecorator(msgs...)
+	Expect(err).To(BeNil())
+
+	// Expect delivery to succeed, then compare balances
+	_, err = testutil.DeliverEthTx(s.ctx, s.app, nil, msgs...)
+	Expect(err).To(BeNil())
+
+	fb := s.app.BankKeeper.GetBalance(s.ctx, funder, denom)
+	db := s.app.BankKeeper.GetBalance(s.ctx, dest, denom)
+
+	s.Require().Equal(funderBalance, fb)
+	s.Require().Equal(destBalance.AddAmount(amount).Amount.Mul(math.NewInt(int64(numTestAccounts))), db.Amount)
+
+	for i, account := range testAccounts {
+		gb := s.app.BankKeeper.GetBalance(s.ctx, account.address, denom)
+		// Use GreaterOrEqual because the gas fee is non-recoverable
+		s.Require().GreaterOrEqual(granteeBalances[i].SubAmount(amount).Amount.Uint64(), gb.Amount.Uint64())
+	}
+}
+
+// delegate is a helper function which creates a message to delegate a given amount of tokens
+// to a validator and checks if the Cosmos vesting delegation decorator returns no error.
+func delegate(clawbackAccount *types.ClawbackVestingAccount, amount math.Int) error {
+	addr, err := sdk.AccAddressFromBech32(clawbackAccount.Address)
+	s.Require().NoError(err)
+
+	val, err := sdk.ValAddressFromBech32("evmosvaloper1z3t55m0l9h0eupuz3dp5t5cypyv674jjn4d6nn")
+	s.Require().NoError(err)
+	delegateMsg := stakingtypes.NewMsgDelegate(addr, val, sdk.NewCoin(utils.BaseDenom, amount))
+
+	dec := cosmosante.NewVestingDelegationDecorator(s.app.AccountKeeper, s.app.StakingKeeper, types.ModuleCdc)
+	err = testutil.ValidateAnteForMsgs(s.ctx, dec, delegateMsg)
+	return err
+}
+
+// validateEthVestingTransactionDecorator is a helper function to execute the eth vesting transaction decorator
+// with 1 or more given messages and return any occurring error.
+func validateEthVestingTransactionDecorator(msgs ...sdk.Msg) error {
+	dec := evmante.NewEthVestingTransactionDecorator(s.app.AccountKeeper, s.app.BankKeeper, s.app.EvmKeeper)
+	err = testutil.ValidateAnteForMsgs(s.ctx, dec, msgs...)
+	return err
+}